requirements.txt
# Core dependencies
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
requests
python-dotenv